pyproject.toml
[project]
name = "hackernews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.31.0",
    "lxml[html-clean]>=6.0.1",
    "newspaper3k>=0.2.8",
    "requests>=2.32.5",
    "selenium>=4.35.0",
    "webdriver-manager>=4.0.2",
]
